Create new wgts.common package and move tools.py there
