+++ src/lib/djerba/plugins/wgts/snv_indel/html.py
@@ -2,7 +2,7 @@
 
 import djerba.core.constants as core_constants
 import djerba.plugins.wgts.snv_indel.constants as sic
-from djerba.plugins.wgts.tools import wgts_tools
+from djerba.plugins.wgts.common.tools import wgts_tools
 from djerba.util.html import html_builder as hb
 
 class snv_indel_table_builder:
